Spelling errors / footer / navbar
Corrected some spelling errors;
I created a footer;
I created a navigation bar;

The navigation bar is working fine at a minimum width of 1024px. As it needs to be responsive, I want it to appear filling the entire screen on mobile screens, it will be enabled by a button that is already in the code.